Move GLUT initialization logic into separate .cpp file
